Completed chores adjustment and cleanup
