+++ src/pages/AssignChore.js
@@ -6,11 +6,27 @@ import EmptyBigCal from "../assets/images/chores/EmptyBigCal.png";
 import statusbar from "../assets/images/statusbar.png";
 import Navbar from "../components/NavbarChoresBlue";
 import AssignButton from "../assets/images/chores/AssignButton.png";
-import React from "react";
+import ChoreBigCalendar from "../assets/images/chores/ChoreBigCalendar.png";
+import React, { useState } from "react";
+import PopUp from "../components/AssignedPopUp";
 
 const AssignChore = () => {
+  //State variables
+  const [modalActive, setModalActive] = useState(false);
+  const [popUp, setPopUp] = useState(false);
+  const [calClicked, setCalClicked] = useState(false);
+
+  //Functions
+  const showChoresModal = () => {
+    setPopUp(true);
+    setModalActive(true);
+  };
+
   return (
     <div className="App">
+      <span className={modalActive ? "modal-background" : ""}>
+        {popUp && <PopUp />}
+      </span>
       <img alt="status bar for phone" src={statusbar} />
       <div className="Single-Icon-Header">
         <div className="Single-Icon-Header-Font">Assign Chore</div>
@@ -35,10 +51,28 @@ const AssignChore = () => {
         name="fname"
       ></input>
       <p>Complete By</p>
-      <Link to="/AssignChoreCalUpdate">
-        <img alt="Chore Calendar Big" src={EmptyBigCal} />
-      </Link>
+      {!calClicked && (
+        <button
+          alt="click to assign chores on a date"
+          onClick={setCalClicked}
+          className="chores-calendar-button"
+        >
+          <img
+            alt="Chore Calendar Big"
+            src={EmptyBigCal}
+            className="chores-calendar"
+          />
+        </button>
+      )}
+      {calClicked && (
+        <img
+          alt="Chore Calendar Big"
+          src={ChoreBigCalendar}
+          className="chores-calendar"
+        />
+      )}
       <img
+        onClick={calClicked && showChoresModal}
         alt="assign button cannot be pressed until you finish filling the form"
         src={AssignButton}
       />
